improve structure of signals
