+++ loc_insurancemaps/models.py
@@ -10,27 +10,19 @@ from pygments import highlight
 from pygments.formatters.html import HtmlFormatter
 from pygments.lexers.data import JsonLexer, JsonLdLexer
 
-
 from django.conf import settings
 from django.contrib.postgres.fields import JSONField
 from django.core.files import File
 from django.core.files.base import ContentFile
 from django.db import models, transaction
-from django.db.models import signals
-from django.dispatch import receiver
 from django.utils.safestring import mark_safe
 from django.urls import reverse
 
 from geonode.base.models import Region, License
-from geonode.layers.models import Layer
-from geonode.documents.models import (
-    Document,
-    post_save_document,
-)
+from geonode.documents.models import Document
 from geonode.documents.renderers import generate_thumbnail_content
 from geonode.people.models import Profile
 
-from georeference.models import LayerMask
 from georeference.proxy_models import DocumentProxy, LayerProxy
 from georeference.utils import TKeywordManager
 
@@ -503,60 +495,3 @@ class Volume(models.Model):
             "urls": self.get_urls(),
             "ordered_layers": ordered_layers,
         }
-
-# connect the creation of a FullThumbnail to Document post_save signal
-@receiver(signals.post_save, sender=Document)
-def create_full_thumbnail(sender, instance, **kwargs):
-    if bool(instance.doc_file) is False:
-        return
-    if not FullThumbnail.objects.filter(document=instance).exists():
-        thumb = FullThumbnail(document=instance)
-        thumb.save()
-
-# triggered whenever a tkeyword is changed on a Document or Layer
-@receiver(signals.m2m_changed, sender=Document.tkeywords.through)
-@receiver(signals.m2m_changed, sender=Layer.tkeywords.through)
-def resource_status_changed(sender, instance, action, **kwargs):
-    """Trigger the document_lookup and layer_lookup updates on a volume
-    whenever a Document or Layer has its georeferencing status changed."""
-    volume = None
-    if action == "post_add":
-        new_status = TKeywordManager().get_status(instance)
-        if new_status is not None:
-            model_name = instance._meta.model.__name__
-            if model_name == "Document":
-                volume = get_volume("document", instance.pk)
-                if volume is not None:
-                    volume.update_document_lookup(instance.pk)
-            if model_name == "Layer":
-                volume = get_volume("layer", instance.pk)
-                if volume is not None:
-                    volume.update_layer_lookup(instance.alternate)
-
-# refresh the lookup for a layer after it is saved.
-@receiver(signals.post_save, sender=Layer)
-def refresh_layer_lookup(sender, instance, **kwargs):
-    volume = get_volume("layer", instance.pk)
-    if volume is not None:
-        volume.update_layer_lookup(instance.alternate)
-
-# pre_delete, remove the reference to the layer in Volume lookups
-# refresh the lookup for a layer after it is saved.
-@receiver(signals.pre_delete, sender=Layer)
-def remove_layer_from_lookup(sender, instance, **kwargs):
-    volume = get_volume("layer", instance.pk)
-    if volume is not None:
-        if instance.alternate in volume.layer_lookup:
-            del volume.layer_lookup[instance.alternate]
-        if instance.alternate in volume.ordered_layers['layers']:
-            volume.ordered_layers['layers'].remove(instance.alternate)
-        if instance.alternate in volume.ordered_layers['index_layers']:
-            volume.ordered_layers['index_layers'].remove(instance.alternate)
-        volume.save(update_fields=["layer_lookup", "ordered_layers"])
-
-# refresh the layer lookup after a LayerMask is deleted.
-@receiver(signals.post_delete, sender=LayerMask)
-def refresh_volume_on_layermask_delete(sender, instance, **kwargs):
-    volume = get_volume("layer", instance.layer.pk)
-    if volume is not None:
-        volume.update_layer_lookup(instance.layer.alternate)
